requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
jsonschema==4.16.0
oauthlib==3.2.2
phonenumbers==8.12.57
psycopg2-binary==2.9.4
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
twilio==7.15.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
